+++ frontend/containers/UploadFileForm/index.tsx
@@ -1,0 +1,3 @@
+import UploadFileForm from "./UploadFileForm";
+
+export { UploadFileForm };
