+++ frontend/pages/workspaces/[workspaceId]/folders/[folderId]/upload-file.tsx
@@ -1,22 +1,15 @@
-import React, { useState } from "react";
+import React from "react";
 
-import { BlockBlobClient } from "@azure/storage-blob";
 import { NextPage } from "next";
 import { useRouter } from "next/dist/client/router";
-import { Input, Form, Button } from "nhsuk-react-components";
-import { useForm, useFieldArray } from "react-hook-form/dist/index.ie11";
 import styled from "styled-components";
-import { Client } from "urql";
 
 import { MainHeading } from "../../../../../components/MainHeading";
 import { NavHeader } from "../../../../../components/NavHeader";
 import { Navigation } from "../../../../../components/Navigation";
 import { PageLayout } from "../../../../../components/PageLayout";
-import { StatusTag } from "../../../../../components/StatusTag";
-import { Textarea } from "../../../../../components/Textarea";
+import { UploadFileForm } from "../../../../../containers/UploadFileForm";
 import {
-  FileUploadUrlsDocument,
-  useCreateFileMutation,
   useGetFolderByIdQuery,
   useGetWorkspaceByIdQuery,
 } from "../../../../../lib/generated/graphql";
@@ -24,10 +17,6 @@ import withUrqlClient from "../../../../../lib/withUrqlClient";
 
 const ContentWrapper = styled.div`
   display: flex;
-`;
-
-const StyledInput = styled(Input)`
-  border: none;
 `;
 
 const PageContent = styled.div`
@@ -43,62 +32,10 @@ const PageContent = styled.div`
   `}
 `;
 
-const StyledButton = styled(Button)`
-  margin-left: 12px;
-`;
-
-const StyledFileInfoBox = styled.div`
-  ${({ theme }) => `
-background-color: ${theme.colorNhsukGrey4};
-`}
-  margin-bottom: 24px;
-`;
-
-const FormField = styled.div`
-  padding: 16px 20px;
-
-  #text {
-    padding-bottom: 60px;
-  }
-`;
-
-const StyledHeadingSection = styled.div`
-  display: flex;
-  direction: column;
-  justify-content: space-between;
-`;
-
-const StyledFileName = styled.h4`
-  width: 225px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-`;
-
-const MAX_CHARS: { [key: string]: number } = {
-  title: 50,
-  description: 250,
-};
-
-interface FormData {
-  files: FileList;
-  fileData: Array<{ title: string; description: string }>;
-}
-
-const UploadFile: NextPage<any> = ({ urqlClient }: { urqlClient: Client }) => {
+const UploadFile: NextPage<any> = () => {
   const router = useRouter();
   const workspaceId = (router.query.workspaceId || "unknown").toString();
   const folderId = (router.query.folderId || "unknown").toString();
-
-  const {
-    register,
-    handleSubmit,
-    errors,
-    setError,
-    control,
-    setValue,
-  } = useForm<FormData>();
 
   const [workspace] = useGetWorkspaceByIdQuery({
     variables: { id: workspaceId },
@@ -106,7 +43,6 @@ const UploadFile: NextPage<any> = ({ urqlClient }: { urqlClient: Client }) => {
   const [folder] = useGetFolderByIdQuery({
     variables: { id: folderId },
   });
-  const [, createFile] = useCreateFileMutation();
 
   if (workspace.error || folder.error)
     return (
@@ -115,133 +51,6 @@ const UploadFile: NextPage<any> = ({ urqlClient }: { urqlClient: Client }) => {
         Oh no... {workspace.error?.message} {folder.error?.message}{" "}
       </p>
     );
-
-  const backToPreviousPage = () => router.back();
-
-  const onSubmit = async (formData: FormData) => {
-    try {
-      const { error, data } = await urqlClient
-        .query(FileUploadUrlsDocument, { count: formData.files.length })
-        .toPromise();
-      if (error) {
-        throw new Error(`Failed to get upload URL: ${error.toString()}`);
-      }
-
-      if (data) {
-        const results: Array<boolean> = await Promise.all(
-          data.fileUploadUrls.map(
-            async (url: string, index: number): Promise<boolean> => {
-              try {
-                const blobClient = new BlockBlobClient(url);
-                const uploadResponse = await blobClient.uploadBrowserData(
-                  formData.files[index],
-                  {
-                    maxSingleShotSize: 4 * 1024 * 1024,
-                  }
-                );
-
-                if (uploadResponse.errorCode) {
-                  console.log(uploadResponse.errorCode);
-                  throw new Error(
-                    `Failed to upload file: ${uploadResponse.errorCode}`
-                  );
-                }
-                const { name: fileName, type: fileType } = formData.files[
-                  index
-                ];
-                const newTitle = formData.fileData[index].title;
-
-                const setMetaResponse = await blobClient.setMetadata({
-                  newTitle,
-                  fileName,
-                });
-
-                if (setMetaResponse.errorCode) {
-                  throw new Error(
-                    `Failed to set file metadata: ${setMetaResponse.errorCode}`
-                  );
-                }
-                if (newTitle === "title") {
-                  throw new Error("FAKE ERROR");
-                }
-                const file = await createFile({
-                  newFile: {
-                    description: formData.fileData[index].description,
-                    fileName,
-                    fileType,
-                    folder: folderId,
-                    temporaryBlobStoragePath: url,
-                    title: newTitle,
-                  },
-                });
-
-                if (file.error) {
-                  console.log("******file error", file.error);
-                  throw new Error(
-                    `Failed to save file: ${file.error?.message}`
-                  );
-                }
-                return true;
-              } catch (error) {
-                setError(`fileData[${index}].title`, {
-                  type: "server",
-                  message: error.toString(),
-                });
-                return false;
-              }
-            }
-          )
-        );
-        if (results.every(Boolean)) {
-          router.push(`/workspaces/${workspaceId}/folders/${folderId}`);
-        } else {
-          console.log("RESULTS", results);
-          setResults(results);
-        }
-      }
-    } catch (error) {
-      setError("files", {
-        type: "server",
-        message: error.toString(),
-      });
-    }
-  };
-
-  const [results, setResults] = useState<Array<Boolean>>([]);
-
-  const { fields } = useFieldArray({
-    control,
-    name: "fileData",
-  });
-
-  const [remainingChars, setRemainingChars] = useState<{
-    [key: string]: number;
-  }>({});
-
-  const handleRemainingChars = (fieldType: "title" | "description") => (
-    event: React.FormEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => {
-    const maxChars = MAX_CHARS[fieldType];
-    setRemainingChars({
-      ...remainingChars,
-      [event.currentTarget.name]: maxChars - event.currentTarget.value.length,
-    });
-  };
-
-  const handleFiles = (files: FileList | null) => {
-    if (files === null) {
-      return;
-    }
-    const filenames = Object.values(files).map((file) => file.name); //['name1', 'name2']
-
-    setValue(
-      "fileData",
-      filenames.map((name) => ({
-        title: name,
-        description: "",
-      }))
-    );
-  };
 
   return (
     <PageLayout>
@@ -263,95 +72,7 @@ const UploadFile: NextPage<any> = ({ urqlClient }: { urqlClient: Client }) => {
               : folder.data?.folder.title || "No title!"}
           </MainHeading>
           <p> Fields marked with * are mandatory.</p>
-          <Form id="filesUploadForm" onSubmit={handleSubmit(onSubmit)}>
-            <StyledInput
-              type="file"
-              name="files"
-              id="files"
-              hint="Maximum 5 files"
-              multiple
-              onChange={(e) => handleFiles(e.currentTarget.files)}
-              inputRef={register({
-                required: {
-                  value: true,
-                  message: "Please select one or more files",
-                },
-              })}
-              aria-invalid={errors.files ? "true" : "false"}
-              error={errors.files?.message}
-            />
-
-            <p>
-              All uploaded content must conform to the platform&apos;s{" "}
-              <a href="#">Terms and Conditions</a>.
-            </p>
-            {fields.map((item, index) => {
-              return (
-                <StyledFileInfoBox key={item.id}>
-                  <FormField>
-                    <StyledHeadingSection>
-                      <StyledFileName>{item.title}</StyledFileName>
-                      {results.length > 0 && (
-                        <StatusTag
-                          successStatus={results[index]}
-                          successMessage="UPLOADED"
-                          failedMessage="UPLOAD FAILED"
-                        />
-                      )}
-                    </StyledHeadingSection>
-                    <Input
-                      type="text"
-                      name={`fileData[${index}].title`}
-                      defaultValue={item.title}
-                      label="Enter file title*"
-                      hint="The title of your file should accurately reflect its content or audience"
-                      onChange={handleRemainingChars("title")}
-                      inputRef={register({
-                        required: {
-                          value: true,
-                          message: "Title is required",
-                        },
-                        maxLength: {
-                          value: MAX_CHARS.title,
-                          message: `Title must be a maximum of ${MAX_CHARS.title} characters`,
-                        },
-                      })}
-                      error={errors.fileData?.[index]?.title?.message}
-                    />
-                    {`${
-                      remainingChars[`fileData[${index}].title`] ||
-                      MAX_CHARS.title
-                    } characters remaining`}
-                  </FormField>
-                  <FormField>
-                    <Textarea
-                      name={`fileData[${index}].description`}
-                      label="Enter description (optional)"
-                      error={errors.fileData?.[index]?.description?.message}
-                      hint="This is the description as seen by users"
-                      onChange={handleRemainingChars("description")}
-                      inputRef={register({
-                        maxLength: {
-                          value: MAX_CHARS.description,
-                          message: `Description must be a maximum of ${MAX_CHARS.description} characters`,
-                        },
-                      })}
-                    />
-                    {`${
-                      remainingChars[`fileData[${index}].description`] ||
-                      MAX_CHARS.description
-                    } characters remaining`}
-                  </FormField>
-                </StyledFileInfoBox>
-              );
-            })}
-            <Button type="submit" name="submitButton">
-              Uploads and continue
-            </Button>
-            <StyledButton secondary type="button" onClick={backToPreviousPage}>
-              Discard
-            </StyledButton>
-          </Form>
+          <UploadFileForm workspaceId={workspaceId} folderId={folderId} />
         </PageContent>
       </ContentWrapper>
     </PageLayout>
